+++ app/policies/core_data_connector/project_policy.rb
@@ -62,7 +62,7 @@ module CoreDataConnector
 
     # Allowed create/update attributes.
     def permitted_attributes
-      [:name, :description, :discoverable, :faircopy_cloud_url]
+      [:name, :description, :discoverable, :faircopy_cloud_url, :faircopy_cloud_project_model_id]
     end
 
     private
