--- db/migrate/20240326171810_add_fcc_import_to_project_models.rb
@@ -1,5 +1,0 @@
-class AddFccImportToProjectModels < ActiveRecord::Migration[7.0]
-  def change
-    add_column :core_data_connector_project_models, :allow_fcc_import, :boolean
-  end
-end
