--- db/migrate/20240326175454_add_fcc_url_to_projects.rb
@@ -1,5 +1,0 @@
-class AddFccUrlToProjects < ActiveRecord::Migration[7.0]
-  def change
-    add_column :core_data_connector_projects, :faircopy_cloud_url, :string
-  end
-end
